+++ enums/siteMap.ts
@@ -1,3 +1,4 @@
+
 export enum SiteMap {
 	HOME = "/",
 	TALK_TO_A_BRAIN = "/talkToABrain",